requirements.txt
Django==1.5.1
Pillow==2.0.0
South==0.7.6
argparse==1.2.1
distribute==0.6.34
django-breadcrumbs==1.1.3-p1
django-ckeditor==4.0.2
pytils==0.2.3
sorl-thumbnail==11.12
wsgiref==0.1.2
